Desugar lib Time API conversions
- Add Time exhaustive tests
- Add API conversion testing option
- Fix double parameter type to convert
- Fix Duration stub

Bug: 134732760

+++ src/test/desugaredLibraryConversions/stubs/Duration.java
@@ -9,7 +9,7 @@ public class Duration {
     return null;
   }
 
-  public int getSeconds() {
+  public long getSeconds() {
     return 0;
   }
